solar_tracking/pyproject.toml
[project]
name = "solar-tracking"
version = "0.1.0"
description = ""
authors = [
    {name = "Akhila-06",email = "akhilachegondi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "awsiotpythonsdk (>=1.5.4,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
